Creating empty classes
