+++ pdf_generator.py
@@ -1,0 +1,10 @@
+import os
+
+
+class PdfGenerator:
+
+    def __init__(self, ingredients):
+        self.ingredients = ingredients
+
+    def create(self):
+        pass
